fix: address SQLite query plan review feedback
